+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# [4.1.3]
+* Fix selection handles show/hide after paste, backspace, copy.
+
+# [4.1.2]
+* Add full support for hardware keyboards (Chromebook, Android tablets, etc) that don't alter screen UI.
+
 # [4.1.1]
 * Added textSelectionControls field in QuillEditor.